+++ packages/warriorjs-abilities/src/look.js
@@ -4,7 +4,7 @@ const defaultDirection = FORWARD;
 
 function look({ range }) {
   return unit => ({
-    description: `Returns an array of up to ${range} spaces in the given direction (${defaultDirection} by default).`,
+    description: `Returns an array of up to ${range} spaces in the given direction (\`'${defaultDirection}'\` by default).`,
     perform(direction = defaultDirection) {
       const offsets = Array.from(new Array(range), (_, index) => index + 1);
       const spaces = offsets.map(offset =>