+++ packages/warriorjs-abilities/src/maxHealth.js
@@ -1,0 +1,10 @@
+function maxHealth() {
+  return unit => ({
+    description: 'Returns an integer representing your maximum health.',
+    perform() {
+      return unit.maxHealth;
+    },
+  });
+}
+
+export default maxHealth;