+++ packages/warriorjs-cli/src/ui/printLevelReport.test.js
@@ -15,24 +15,22 @@ const profile = {
 };
 
 const score = {
-  parts: {
-    warrior: 3,
-    timeBonus: 2,
-    clearBonus: 0,
-  },
-  total: 5,
-  grade: 0.9,
+  warrior: 3,
+  timeBonus: 2,
+  clearBonus: 0,
 };
+const totalScore = 5;
+const grade = 0.9;
 
 test('prints level score', () => {
-  printLevelReport(profile, score);
+  printLevelReport(profile, score, totalScore, grade);
   expect(printLine).toHaveBeenCalledWith('Warrior Score: 3');
   expect(printLine).toHaveBeenCalledWith('Time Bonus: 2');
   expect(printLine).toHaveBeenCalledWith('Clear Bonus: 0');
 });
 
 test('prints regular score if not epic', () => {
-  printLevelReport(profile, score);
+  printLevelReport(profile, score, totalScore, grade);
   expect(printTotalScore).toHaveBeenCalledWith(3, 5);
 });
 
@@ -42,13 +40,13 @@ describe('epic', () => {
   });
 
   test('prints epic score', () => {
-    printLevelReport(profile, score);
+    printLevelReport(profile, score, totalScore, grade);
     expect(printTotalScore).toHaveBeenCalledWith(2, 5);
   });
 
   test('prints level grade', () => {
     getGradeLetter.mockReturnValue('A');
-    printLevelReport(profile, score);
+    printLevelReport(profile, score, totalScore, grade);
     expect(printLine).toHaveBeenCalledWith('Level Grade: A');
     expect(getGradeLetter).toHaveBeenCalledWith(0.9);
   });