+++ CHANGELOG.md
@@ -8,6 +8,42 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.14.0] - 2018-10-17
+
+### Changed
+
+- Simplify `@warriorjs/helper-get-level-score` ([@olistic][] in [#247][])
+- Rename `@warriorjs/helper-get-play-score` package to
+  `@warriorjs/helper-get-level-score` ([@olistic][] in [#246][])
+
+## [0.13.0] - 2018-10-09
+
+### Added
+
+- Add `maxHealth` ability ([@jseed][] in [#238][])
+
+### Changed
+
+- Improve abilities descriptions ([@olistic][] in [#240][])
+
+## [0.12.3] - 2018-10-04
+
+### Fixed
+
+- Fix `getLevelConfig` (make it pure for real) ([@olistic][] in [#236][])
+
+## [0.12.2] - 2018-10-04
+
+### Fixed
+
+- Make `getLevelConfig` a pure function ([@olistic][] in [#234][])
+
+## [0.12.1] - 2018-07-31
+
+### Fixed
+
+- Fix `tick-tick-boom` tower description
+
 ## [0.12.0] - 2018-07-30
 
 ### Changed
@@ -205,7 +241,12 @@ whose logic can be reused by different flavors of the game.
 
 Initial version.
 
-[unreleased]: https://github.com/olistic/warriorjs/compare/v0.12.0...HEAD
+[unreleased]: https://github.com/olistic/warriorjs/compare/v0.14.0...HEAD
+[0.14.0]: https://github.com/olistic/warriorjs/compare/v0.13.0...v0.14.0
+[0.13.0]: https://github.com/olistic/warriorjs/compare/v0.12.3...v0.13.0
+[0.12.3]: https://github.com/olistic/warriorjs/compare/v0.12.2...v0.12.3
+[0.12.2]: https://github.com/olistic/warriorjs/compare/v0.12.1...v0.12.2
+[0.12.1]: https://github.com/olistic/warriorjs/compare/v0.12.0...v0.12.1
 [0.12.0]: https://github.com/olistic/warriorjs/compare/v0.11.3...v0.12.0
 [0.11.3]: https://github.com/olistic/warriorjs/compare/v0.11.2...v0.11.3
 [0.11.2]: https://github.com/olistic/warriorjs/compare/v0.11.1...v0.11.2
@@ -235,6 +276,7 @@ Initial version.
 [@pigalot]: https://github.com/pigalot
 [@djohalo2]: https://github.com/djohalo2
 [@wtgtybhertgeghgtwtg]: https://github.com/wtgtybhertgeghgtwtg
+[@jseed]: https://github.com/JSeed
 [#67]: https://github.com/olistic/warriorjs/pull/67
 [#70]: https://github.com/olistic/warriorjs/pull/70
 [#77]: https://github.com/olistic/warriorjs/pull/77
@@ -285,3 +327,9 @@ Initial version.
 [#219]: https://github.com/olistic/warriorjs/pull/219
 [#220]: https://github.com/olistic/warriorjs/pull/220
 [#221]: https://github.com/olistic/warriorjs/pull/221
+[#234]: https://github.com/olistic/warriorjs/pull/234
+[#236]: https://github.com/olistic/warriorjs/pull/236
+[#238]: https://github.com/olistic/warriorjs/pull/238
+[#240]: https://github.com/olistic/warriorjs/pull/240
+[#246]: https://github.com/olistic/warriorjs/pull/246
+[#247]: https://github.com/olistic/warriorjs/pull/247